requirements.txt
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.9.0
authlib>=1.2.0
flask>=2.3.0

